fix(suite-native): navigation hooks refactor
