+++ suite-native/module-authorize-device/src/screens/passphrase/PassphraseEnterOnTrezorScreen.tsx
@@ -1,14 +1,5 @@
-import { useCallback } from 'react';
-import { useSelector } from 'react-redux';
-
-import { useFocusEffect } from '@react-navigation/native';
-
 import { EventType } from '@suite-native/analytics';
 import { Text } from '@suite-native/atoms';
-import {
-    DeviceAuthorizationStep,
-    selectDeviceAuthorizationStep,
-} from '@suite-native/device-authorization';
 import { Translation } from '@suite-native/intl';
 import { AuthorizeDeviceStackRoutes, useNavigateToInitialScreen } from '@suite-native/navigation';
 import {
@@ -18,19 +9,13 @@ import {
 import { useLegacyAnalytics } from '@suite-native/services';
 import TrezorConnect from '@trezor/connect';
 
+import { useHandleNavigateToInitialScreenOnIdle } from '../../hooks/useHandleNavigateToInitialScreenOnIdle';
+
 export const PassphraseEnterOnTrezorScreen = () => {
     const navigateToInitialScreen = useNavigateToInitialScreen();
     const legacyAnalytics = useLegacyAnalytics();
-    const deviceAuthorizationStep = useSelector(selectDeviceAuthorizationStep);
 
-    useFocusEffect(
-        useCallback(() => {
-            if (deviceAuthorizationStep === DeviceAuthorizationStep.Idle) {
-                // NOTE: this means that the device passphrase request was cancelled on the device so we need to navigate back
-                navigateToInitialScreen();
-            }
-        }, [deviceAuthorizationStep, navigateToInitialScreen]),
-    );
+    useHandleNavigateToInitialScreenOnIdle();
 
     const handleCancel = () => {
         legacyAnalytics.report({
