+++ suite-native/passphrase/src/index.ts
@@ -11,4 +11,4 @@ export * from './components/EmptyWalletInfoSheet';
 export * from './components/EnterPassphraseOnTrezorButton';
 export * from './components/NoPassphraseButton';
 export * from './passphraseSelectors';
-export * from './useWaitForUiRequestPassphraseOnDevice';
+export * from './useHandleUiRequestPassphraseOnDevice';
