+++ suite-native/passphrase/src/useHandleUiRequestPassphraseOnDevice.ts
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 
 import TrezorConnect, { UI } from '@trezor/connect';
 
-export const useWaitForUiRequestPassphraseOnDevice = (callback: () => void) => {
+export const useHandleUiRequestPassphraseOnDevice = (callback: () => void) => {
     useEffect(() => {
         TrezorConnect.on(UI.REQUEST_PASSPHRASE_ON_DEVICE, callback);
